requirements.txt
# Requirements for improved Version System

# Core dependencies
GitPython==3.1.37
PyYAML==6.0.1
dataclasses-json==0.6.3

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional dependencies for enhanced features
click==8.1.7  # For better CLI experience
rich==13.7.0  # For enhanced terminal output
pydantic==2.5.0  # For configuration validation

# Security
bandit==1.7.5  # Security linting

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
